# Vectorized the util calculations